+++ 3D/Incompressible/Laminar/FSM/Explicit/ForwardEuler/src/IO.cc
@@ -2,7 +2,7 @@
 #include "../inc/custom_type.h"
 #include "../inc/IO.h"
 
-extern size_t NumOfPnt, NumOfFace, NumOfCell;
+extern int NumOfPnt, NumOfFace, NumOfCell;
 extern NaturalArray<Point> pnt;
 extern NaturalArray<Face> face;
 extern NaturalArray<Cell> cell;
